# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "اغتنم API Documentation",
        "description": "API documentation for اغتنم application",
        "contact": {
            "name": "Admin",
            "email": "admin@e8tanem.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/podcasts": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "List all podcasts",
                "description": "Returns a paginated list of active podcasts with pagination metadata",
                "operationId": "74471147bb5a65d449b5357445399019",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination (starts from 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 10)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter podcasts by title or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Field to sort by (created_at, plays_count)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "created_at",
                            "enum": [
                                "created_at",
                                "plays_count"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of podcasts with pagination metadata",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Sample Podcast"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Podcast description"
                                                    },
                                                    "audio_url": {
                                                        "type": "string",
                                                        "example": "/storage/podcasts/sample-podcast.mp3"
                                                    },
                                                    "thumbnail_url": {
                                                        "type": "string",
                                                        "example": "/storage/thumbnails/podcast-thumbnail.jpg"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Technology"
                                                    },
                                                    "plays_count": {
                                                        "type": "integer",
                                                        "example": 50
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/podcasts?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/podcasts?page=5"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "&laquo; Previous"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/podcasts?page=2",
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/podcasts"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "http://e8tanem.com/api/podcasts?page=1",
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/podcasts/{podcast}": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get a specific podcast",
                "operationId": "c993f9fc45e1142860519dc9fa92a247",
                "parameters": [
                    {
                        "name": "podcast",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Podcast details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "audio_url": {
                                            "type": "string"
                                        },
                                        "thumbnail_url": {
                                            "type": "string"
                                        },
                                        "category": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "plays_count": {
                                            "type": "integer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Podcast not found"
                    }
                }
            }
        },
        "/api/podcasts/{podcast}/play": {
            "post": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Increment play count for a podcast",
                "operationId": "08a00ec429f95578fe1489ce74f29e49",
                "parameters": [
                    {
                        "name": "podcast",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Play count incremented",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plays_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/podcasts/{podcast}/stream": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Get podcast audio file URL",
                "description": "Returns the direct URL to the podcast audio file for regular playback",
                "operationId": "f74b7debb6336d003e4603e337f11161",
                "parameters": [
                    {
                        "name": "podcast",
                        "in": "path",
                        "description": "ID of the podcast to get audio URL for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audio file URL",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "audio_url": {
                                            "type": "string",
                                            "example": "https://example.com/storage/podcasts/sample.mp3"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Podcast Title"
                                        },
                                        "podcast_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Podcast not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}